dom/tests/mochitest/general/test_showModalDialog_removed.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1077002
-->
<head>
  <title>Test for showModalDialog unavailability in e10s</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank"
    href="https://bugzilla.mozilla.org/show_bug.cgi?id=1077002">Mozilla Bug 1077002</a>
<p id="display"></p>
<div id="content">
  <iframe id="frame" style="height:100px; width:100px; border:0"></iframe>
  <div id="status" style="display: none"></div>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for showModalDialog unavailability in Firefox **/

// NB: This test runs in e10s only. In e10s showModalDialog should not
// exist.
ok(!window.showModalDialog, "showModalDialog should not exist");

</script>
</pre>
</body>
</html>
